cleanup: home the pip failure report beside the pip machinery
The report family encodes a fact about the pip subprocess (which stream
carries which log level), so it moves from update to platform::process,
next to pip_command and tail_for_log; update now calls down through
platform::pip_output_report instead of platform reaching up into update.
The single-caller stderr_tail closure is gone, the truncation marker no
longer claims everything it truncates is stderr, the stub-interpreter
recipe is hoisted to module level in python_env so both test mods share
it, and the stub test is trimmed to the wiring assertions the pure
fixture tests cannot cover.
